renamed StudentDetailsResource to StudentDetailsScreen
